[REFACTOR] 트랜잭션 수정
