+++ src/test/java/kr/hhplus/be/server/interfaces/api/pointhistory/PointHistoryResponse.java
@@ -1,3 +1,3 @@
-package kr.hhplus.be.server.interfaces.api.point_history;
+package kr.hhplus.be.server.interfaces.api.pointhistory;
 
 public record PointHistoryResponse(Long userId, int amount, String type) {}
